lutfi-project-master/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
ffprobe==0.5
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gTTS==2.5.4
h5py==3.13.0
idna==3.10
joblib==1.5.0
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
protobuf==5.29.4
PyAudio==0.2.14
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pyserial==3.5
regex==2024.11.6
requests==2.32.3
rich==14.0.0
setuptools==80.7.1
six==1.17.0
SpeechRecognition==3.14.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
fastapi
uvicorn
fastapi[standard]